Add files via upload
COVID-19 Data Analysis and Forecasting model - it uses matplotlib, pandas in-built functions, plotly and plotly express for data visualization. The World analysis and forecasting have been done using Fbprophet, Prophet is a procedure for forecasting time-series data based on the additive model where non-linear trends are fit with yearly, weekly and daily seasonality. It is robust to outliers and other missing data.