Create README.md
COVID-19 Data Analysis and Forecasting - Uses matplotlib, pandas in-built functions, plotly and plotly express for data visualization and for the forecasting Fbprophet has been imported. Prophet is a procedure for forecasting time-series data based on the additive model where non-linear trends are fit with yearly, weekly and daily seasonality. It is robust to outliers and other missing data.